pkg/homedir: enhance GetConfigHome()
copy the version from Podman that checks if the
user owns the directory.

[NO NEW TESTS NEEDED] moved from Podman
